Move connect modal provider to context
