+++ src/lib/wallets/useConnectModal.tsx
@@ -1,19 +1,1 @@
-import { createContext, useContext } from "react";
-
-type ConnectModalContextValue = {
-  openConnectModal: (() => void) | undefined;
-  connectModalOpen: boolean;
-};
-
-export const ConnectModalContext = createContext<ConnectModalContextValue>({
-  openConnectModal: undefined,
-  connectModalOpen: false,
-});
-
-/**
- * Drop-in replacement for RainbowKit's useConnectModal.
- * Returns { openConnectModal, connectModalOpen } with the same interface.
- */
-export function useConnectModal() {
-  return useContext(ConnectModalContext);
-}
+export { useConnectModal } from "context/ConnectModalContext/ConnectModalContext";
